chore: Isolating tests with exec dependency
